requirements.txt
langchain>=0.3.0
langchain-openai>=0.2.0
langgraph>=0.2.10
supabase>=2.7.0
vecs>=0.4.0
postgrest>=0.18.0
mcp>=1.0.0
httpx>=0.27.0
sse-starlette>=2.1.0
streamlit>=1.38.0
python-dotenv>=1.0.1
pydantic>=2.9.0
